Produce output that is in canonical form.
Magnitude nonnegative, phase within (-pi,pi], (magnitude == 0) => (phase == 0).

To make this happen, we need to replace the exported data constructor with a
pattern synonym wrapping the smart constructor.

On GHC>=7.8.6, (:<) is now a pattern synonym, effectively an alias for mkPolar.
On Hugs and GHC<7.8.6, the symbol (:<) is undefined.

The data constructor of Polar is renamed to (:<<), and it is not exported.

The output format of show/read has changed:
    show (1 :< 3.14) = "mkPolar 1 3.14"

There are two obvious alternatives, the first is:
    show (1 :< 3.14) = "1 :< 3.14"
This format is unacceptable because it is not executable code on Hugs.

The second is:
    show (1 :< 3.14) = "1 :<< 3.14"
This format is unacceptable because it exposes the dumb constructor and allows
the construction of non-canonical values via read.

Not tested on GHC <8.0.2, because I can't get it to build.

Also fix the implementation of mkPolar, negate, signum, fromRational, and exp.